LaTeX rendering on the web interface
Integrate tex2png-hs into web-prover and use it for rendering the
assumptions and the goal when clicking Prove.

Notes:

o tex2png-hs uses `latex` and `dvipng` behind the scenes and assumes
that they are installed.

o For the rendering to work, the following four style files are needed
as of now: bsymb.sty, calculation.sty, eventB.sty, unitb.sty

o The tex2png-hs library doesn't check for existence of the sty files,
therefore make sure that they are properly installed before running the
application. See script/install-stylesheets-macosx.hs of the parent
project.

o The LaTeX commands are wrapped in an array environment, which is then
wrapped inside an align* environment.